pyproject.toml
[tool.poetry]
name = "device-readings"
version = "0.1.0"
description = "Device Readings API"
authors = ["Calebe Maciel <calebegeazi@gmail.com>"]
readme = "README.md"
packages = [{include = "device_readings"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "5.1.5"
gunicorn = "23.0.0"
newrelic = "10.3.1"
setuptools = "75.6.0"
psycopg2-binary = "2.9.10"
django-model-utils = "5.0.0"
django-unique-upload = "0.2.1"
djangorestframework = "3.15.2"
boto3 = "1.35.76"
pydantic = "^2.10.5"


[tool.poetry.group.dev.dependencies]
ipdb = "0.13.13"
ipython = "8.30.0"
flake8 = "7.1.1"
mock = "5.1.0"
factory-boy = "3.3.1"
pytest-django = "4.9.0"
coverage = "7.6.9"
drf-yasg = "^1.21.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "device_readings.settings"
python_files = ["*tests.py"]
markers = [
    "unit: standard unit test with no db access"
]